race-car/pyproject.toml
[project]
name = "race-car"
version = "0.1.0"
description = "AI Race Car Challenge"
requires-python = ">=3.11"
dependencies = [
    "pygame>=2.5.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "colorama>=0.4.6",
]

[project.scripts]
api = "api:start_server"
validate = "validate:main"
check-status = "validate:check_status_cli"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
]

docs = [
    "mkdocs>=1.6.1",
    "mkdocs-api-autonav>=0.3.0",
    "mkdocs-autoapi>=0.4.1",
    "mkdocs-awesome-nav>=3.1.2",
    "mkdocs-material>=9.6.16",
    "mkdocstrings>=0.30.0",
    "mkdocstrings-python>=1.16.12",
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true
major_version_zero = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = ["*.py"]
